update minimum and maximum python version
The minimum python version was updated due to dependency conflicts caused by typed-astp from mypy, black and pylint.

The project typed-ast is no longer maintained, and new versions of mypy, pylint and black doe not use it.

Due to the maximum supported version by the project, it was necessary to update the flake8 package. Therefore, it was necessary to make a spacing change in the code of examples/signals.py.